src/dev-requirements.txt
# Copyright (C) 2013 Google Inc., authors, and contributors <see AUTHORS file>
# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
# Created By: dan@reciprocitylabs.com
# Maintained By: dan@reciprocitylabs.com

Jinja2==2.6
# Use `behave` library from github for --junit support of ScenarioOutline
# This will be unnecessary when behave==1.2.2.16 is released to PyPI
# see: https://github.com/behave/behave/issues/69
git+https://github.com/behave/behave.git@1.2.3a19#egg=behave
Flask-Testing
git+https://github.com/rbarrois/factory_boy.git#egg=factory-boy
mock
nose==1.3.0
requests==2.0.0
PyYAML
Flask-Jasmine
Alembic==0.6.4
